requirements.txt
hdx-python-api~=5.9.0
geopandas~=0.10.2
mapbox==0.18.0
numpy~=1.20.2
pandas~=1.3.4
rasterio~=1.2.10
rasterstats~=0.16.0
